api/pyproject.toml
[project]
name = "multi-agent-research-assistant"
version = "1.0.0"
description = "Multi-Agent Research Assistant — search, synthesise, report, evaluate"
requires-python = ">=3.11"
dependencies = [
    "groq>=0.1.0",
    "huggingface-hub>=0.25.0",
    "pydantic>=2.0.0",
    "ddgs>=9.0.0",
    "tavily-python>=0.5.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "markdown>=3.6",
    "weasyprint>=60.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx",  # used by FastAPI TestClient
]

[project.scripts]
research = "main:main"
